Add MerryIntMap; it has quirky performance but does well
It's important to note that MerryIntMap without a pairing function for keys is much faster than IntMap with or without a pairing function for keys. Using a pairing function makes extracting x and y from a key much slower, so it's great that we can avoid it now. MerryIntMap is slightly slower than IntMap when nesting with a Map of Map of V structure, probably because that structure rarely has key collisions and so the extra effort used for Fibonacci hashing is wasted. Fibonacci hashing is what allows the pairing function to be elided, though, for the single-map case.