Further debug Muroz et al's cbrt().
The original version in the paper fails on any negative input (very unlike Math.cbrt()), returning NaN. With the RNG uses here, it fails on the first item. The code here ignores the NaN approximations and replaces them with something almost as inaccurate (-r, which was the input to cbrtProblem(r), just negated). Even with the modifications I made to handle negative inputs, it's still inaccurate over the -512 to 512 range. It might be OK over a smaller range, or may be faster.