Quick fix to make atan2Funky pass SquidLib's tests.
Now, we check if n is not NaN (from when both y and x are infinite) but is infinite, and if it is we set x to be 0 (because y must be infinitely greater than x, and this allows avoiding the infinity-allergic atn() call). This technique has about the same speed, 4% slower, because of branch prediction (the changed code is on an extremely-rare branch). Also, the double version of atan2Funky_ has some mistakes fixed.