Benchmark RomuTrio; quite fast, not number 1.
RomuTrio is faster than Harpo on Java 8, but it's slightly slower on Java 16, and much slower on OpenJ9 15. It is a well-studied and high-quality generator, but some potential flaws are obvious -- like xoroshiro and xoshiro, if it is seeded with all 0 for its full state, it will never produce a result other than 0.