Benchmark a bunch of RNGs on Java 16.
Surprisingly, RandomXS128 is quite fast now, though still slower than the other RNGs that also sped up on Java 16 (Groucho and Harpo). Most other differences were small. It's interesting to note that a generator to be official in Java 17 seems to have no performance gains as of Java 16, and is actually slower on Java 16 than it is on Java 8.