Make (roughly) all of the Strings in the squid MemoryMBeanCheck unique
This has serious effects on the JDK collections at large sizes, since HashMap (and the code that uses it, like HashSet) seemed to suddenly shrink its memory usage for the largest sizes. It doesn't do that in practice; what was happening seemed to be related to duplicate String values or possibly Entry values getting interned or otherwise consolidated. With about 250K Strings possible in the dictionary we use, going from 100K to 1M items started to reuse Strings, but now we use a Set of unique Strings with appended numbers, and HashMap does about the same for 1M as it does 100K or 1K -- not great on memory usage.