Testing Kryo's Cuckoo (with the small fix for stash resizing) vs. libGDX ObjectMaps.
My machine is too hot for absolute speed comparison, but relative comparisons should be fine. Cuckoo seems to be a tiny bit faster on containsKey() most of the time (but much slower in pathological cases), and about the same on iteration, but only has about half the throughput on libGDX's ObjectMap when any insertions are being done (populate and copy). Overall, the current libGDX approach remains better.