Try coolerinc's threaded anim8-gdx, hmm!
It is actually quite a bit faster on Pattern dither, because that one takes so much CPU processing I think. But, it's slower on several others because there's a roughly-2-second minimum time it takes to process one 90-frame GIF, and some dither algorithms take less than 2 seconds normally to do this task.