Funky fresh moves.
This is a neat usage of IntDeque to reduce or eliminate the type of slowdown you see when changing the middle of an IntList. IntList.removeAt() takes O(n) time when not removing from the end of the list, where n is the number of items after the index you're removing. Now we use the O(1) IntDeque.removeLast(), calling it the same number of times, but adding new items to the beginning of the deque (which is O(1) each time) instead of after the index we remove from.