I'm shocked and amazed!
FlipMap actually does just as well with load factor 0.75 as 0.5, which in all the cuckoo hashing information I've found, is supposed to be too high if only using two hash functions. The other hashtables perform worse (jdkgdxds takes almost twice as long) with 0.75 load factor, but FlipMap seems to stay just as good.